third_party/blink/web_tests/external/wpt/shadow-dom/wheel-event-related-target-expected.txt

This is a testharness.js-based test.
[FAIL] wheel relatedTarget should be retargeted to shadow host when entering shadow DOM
  assert_equals: relatedTarget before mouse is over shadow host expected Element node <div id="container">Scrolling container\n  <div id="height... but got null
Harness: the test ran to completion.
